requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.2

# Data Collection
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# NLP and ML
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0

# Topic Modeling
bertopic==0.16.0
sentence-transformers==2.2.2

# Data Processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Time Series Analysis
prophet==1.1.5

# Logging and Utilities
tqdm==4.66.1
loguru==0.7.2

# Development and Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Jupyter Notebooks
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1
